Split out tests and add type test
